refactor(backend): reduce media repository entrypoint surface
